chore(ChartTooltip): Add visual test
This adds a visual test for the case where a title prop is given to the
component, but _not_ a title slot. This is to guard against regressions
as other issues from the [slot migration epic][1] are fixed. In
particular, a naïve fix for [GlPopover][2] would break this behaviour,
so this visual test has been added to guard against that.

[1]: https://gitlab.com/groups/gitlab-org/-/epics/6269
[2]: https://gitlab.com/gitlab-org/gitlab-ui/-/issues/1479