fix(GlSorting): Make dropdown toggle conform to design
This removes a legacy `dropdown-menu-toggle` class that was applied to
the dropdown toggle within the component. That class applied various
irrelevant styles within GitLab itself, but notably also set the width
to a fixed 160px. This violates the design guidelines, disagrees with
the plain GlDropdown's styling, and generally isn't a good idea.