feat: Add GlDisclosure/GlDisclosureItem components
`GlDropdown` component is not very accessible.
That is why we're splitting it into 4 different components
Each one with a better semantics and thus accessibility.
This MR is dedicated to one of the 4 - disclosure component.
This iteration mainly extracts Disclosure and its Item
into separate components copying code from GlDropdown/Item
and removing unneccesary code but NOT adjusting it to meet
accessibility criteria.

More details: https://design.gitlab.com/components/dropdown