fix(GlBadge): Ensure consistent baseline
Previously, the text-only type of `GlBadge` did not share the same
baseline as the icon-only and icon-and-text versions.

This is because `GlBadge`'s icons are set to a height of 16px, whereas
the line-height of the text is 14px. Combined with `.gl-badge`'s styling
of `align-items: center`, this made the *icon* determine the badge's
baseline, rather than the text.

The fix is to make sure that the *text* determines the baseline. For the
icon-and-text case, that's straight forward: set `align-items: baseline`
for the `.gl-badge`, and set the icon to `align-self: center`.

The icon-only case is trickier, since there *is* no text to create
a baseline from. The fix in this case is also simple, but weird: to add
a `.gl-badge::after` pseudo element with the content of a zero-width
space!

It must be a *zero-width* space so that it doesn't affect the rendered
output, other than creating the text baseline if there isn't any other
text.

Addresses https://gitlab.com/gitlab-org/gitlab-ui/-/issues/1673.