ci: run visual tests for changed components only
This adds a script to minimze the amount of visual tests being run in MR
pipelines. The script runs with the `yarn test:visual:minimal` command.

It works iterating through the changed files list to find `.stories.js`
files next to each of them and running storyshots against the inferred
stories only.

If, for any given file, no `.stories.js` file can be found, the change
is considered far-reaching and the full test suite is run.

In any case, the full test suite can be run by triggering the `visual`
CI job which remains available as a manual job in MR pipelines. It
always runs in the default branch.