revert(GlChartSeriesLabel): Add validation for type prop
This reverts
https://gitlab.com/gitlab-org/gitlab-ui/-/merge_requests/2714, which
caused [failures downstream in GitLab][1].

This also highlights that we have missing test coverage in GitLab UI.

It turns out that GlChartLegend passes the series's type (e.g., "bar")
to GlChartSeriesLabel, while GlChartSeriesLabel expects only "solid" or
"dashed".

It's not immediately clear what the correct fix is, hence this revert.

[1]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/83745#note_905812358