feat(GlListbox): convert header slot to a prop
This drops the header slot in favor of a prop. The slot was deemed too
flexible. By only allowing a plain string in the header, we ensure that
listbox usages don't deviate from the designs too easily.

BREAKING CHANGE: GlListbox usages should be updated to pass the header
via the `headerText` prop instead of the `header` slot.