docs(GlDropdown): Document deprecation
This adds a brief note in `GlDropdown`'s documentation that it will be
deprecated in favour of `GlListbox` and other components in future.

An [accessibility audit][audit] determined that `GlDropdown`'s
accessibility semantics are not appropriate for all of the
[use cases][usecases] a dropdown should support.

Once there are enough suitable alternatives, `GlDropdown` should be
marked/renamed as officially deprecated.

See [this epic][epic] for more details.

[audit]: https://gitlab.com/groups/gitlab-org/-/epics/6573
[usecases]: https://gitlab.com/gitlab-org/gitlab-ui/-/issues/1534
[categorise]: https://gitlab.com/gitlab-org/gitlab/-/issues/346804
[epic]: https://gitlab.com/groups/gitlab-org/-/epics/1059