Don't split on spaces
Surprisingly, this apears to be all that is necessary to stop the
special handling of spaces.

What now needs to happen is tokenisation of free text terms, and a way
to "commit" a free text term.

This removes "split" events/handlers, the splitOnQuotes util, and also
makes sure that values bound to the component don't get split on spaces.