Remove unneeded wrapTokenInQuotes util
I don't really understand the circumstances when this is used. For the
milestones in the story with a space, it wraps in quote. For
confidential tokens with spaces, it doesn't. Why?

Ah, is it because the former uses the #suggestions slot, whereas the
latter uses the `options` property on the `available-tokens` prop? Why
are these handled differently?

It's something to do with the quote-wrapping. In the former case,
perhaps those are stripped away? Or maybe that _value_ is used, rather
than the user-facing string?