fix: Use value instead of title in options
The check `[this.optionTextField] === v` *always* evaluates to false,
since the array literal cannot be strictly equal to anything but itself.
So, the optional chaining and nullish coalescing in
` this.getMatchingOptionForInputValue(newValue)?.value ?? newValue`
meant that the title (`newValue`) was used instead of the option's
actual `value`.

BREAKING CHANGE: An option's `value` is now used instead of its `title`
for the chosen suggestion.

This only affects tokens with an `options` property. Tokens which use
the `suggestions` slot are not affected.