Remove check on activeTokenIdx in isLastToken
This negation is truthy when activeTokenIdx is null (there's no active token)
*and* when it's 0 (the first token is active). Should these be handled
distinctly/explicitly/differently?

This means that the last token is told it's *not* the last token if anything
other than the *first* token is active. It's not clear if this is intentional
or not. Could it be this was done to avoid the last token being told it's the
last token if *it* is active?