fix(GlFilteredSearch): Improve activeTokenIdx check in isLastToken
This negation is truthy when activeTokenIdx is null (there's no active
token) *and* when it's 0 (the first token is active).

This means that the last token is told it's *not* the last token if
anything other than the *first* token is active. We don't want to treat
the first token specially like this, so instead we check just for null.