feat(GlDisclosureDropdown): Pass item to action
BREAKING CHANGE: The `this` of the `GlDisclosureDropdown` action
callback is now `undefined` rather than the item it's attached to. It
now receives the item as a regular argument instead.

This makes it more similar to the `action` event that's emitted by the
component, such that the same function could be used in either way.

This also means that the same callback can be attached to all items,
instead of requiring a unique callback for each item.