ci: Create screenshots with `@storybook/test-runner`
Apparently `@storybook/test-runner` is a jest and playwright based test
runner which allows us to test our storybook stories and create
snapshots of them. The benefit of this package over
`@storybook/addon-storyshots`, is that each story becomes it's own test.
That means that jest is able to parallelise tests run with workers.

Furthermore it has a `--shard` parameter to further parallelise tests
across multiple jobs, similar to the playwright parameter with the same
name.

This commit introduces the dependency and utilizes them in new jobs:
`visual` and `update_screenshots`. The old corresponding puppeteer jobs
have been renamed with an `old` prefix.

Future improvements:
- Remove storyshots, it's expectations and everything related
- Enable more browsers (firefox, webkit)
- Change our puppeteer image to become a playwright image
- Consider migrating our few integration tests to playwright