chore(GlDisclosureDropdown): Simplify arg creation
Storybook is already able to infer prop args via the `component`
property of the default export of a story file. There's no need to
explicitly read them ourselves.

This led to a problem for a lazily defined default value in
https://gitlab.com/gitlab-org/gitlab-ui/-/merge_requests/3244#note_1259093614.

It may be this pattern of explicitly reading default values arose before
Storybook inferred component props.