fix(GlTable): ensure gl-table class on table element
The underlying BTable component conditionally wraps the `<table>`
element with a `<div>`. In BootstrapVue 2.20.1, that happens under fewer
circumstances than in Bootstrap 2.23.1, due to
https://github.com/bootstrap-vue/bootstrap-vue/pull/6266.

The reason this causes breakage is that our GlTable component applies
the `gl-table` class to the root element of BTable, and our CSS assumes
that that class will always be applied to a `<table>`. So, under
BootstrapVue 2.23.1, without this fix, there would be more circumstances
where the styling would be incorrect.

This should fix part of the original reason for the revert (see
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/8035).