fix(css): convert gl-sm-pl-0 util to mobile-first
This changes the gl-sm-pl-0 utility mixin to be mobile-first instead of
desktop-first.

Also adds the gl-sm-pl-6 util, necessary to integrate this change in
GitLab.

BREAKING CHANGE: Usages of gl-sm-pl-0 need to be adjusted to account for
the fact it now removes left padding on viewports larger than sm, while
it used to do the opposite.