fix(GlFilteredSearch): Improve handling of some falsey values
This allows `false`, `0` and `''` to be handled better as values in
options. Previously, an option with one of these values could never get
automatically selected when typing their title in a segment.

An example where this can be seen is the Confidential token in Roadmaps:
https://gitlab.com/gitlab-org/gitlab/-/blob/2dcb4f37a26ffff6b2cebc6b1f55f1294b1b545d/ee/app/assets/javascripts/roadmap/mixins/filtered_search_mixin.js#L179-180.

The `null` and `undefined` (and `NaN`) values are still ignored.