fix(GlFilteredSearch): Don't prevent tabbing past component
This is a follow-up to 0798e433c7ac8c22300c579613b9f3aa5bc4b929 which
added support for opening the token suggestion list when tabbing to the
input. Unfortunately, that broke tabbing forward past the component.

This is because the next focusable element in document order after the
input is the first suggestion. The fix therefore is to make it unable to
receive focus via the keyboard.

This doesn't harm accessbility, since that list was never able to
receive focus anyway, since it's removed when the input loses focus.

Ideally there'd be a test for this, but Cypress doesn't support sending
Tab keypresses: https://github.com/cypress-io/cypress/issues/299.