fix(dropdowns): fix ResizeObserver loop error
This rewrites a previous fix that meant to fix an unexpected srolling
issue but ended up causing a ResizeObserver loop error. While the error
is harmless, it does cause an error overlay to show up in the GDK, and
sends events to our Sentry instance.
This applies the simpler fix that was outlined in
https://gitlab.com/gitlab-org/gitlab-ui/-/merge_requests/3484#there-is-a-simpler-approach
which seems to fix both issues.