fix(GlTooltip): Reduce component boundary padding
The tooltip component has a default boundary padding of 50, whereas the
directive has a default boundary padding of 5. There's no obvious reason
for this difference, and it appears to be a bug upstream:
https://github.com/bootstrap-vue/bootstrap-vue/issues/6507.

This bug means that tooltips placed on elements near the configured
boundary (e.g., the viewport) will render unnecessarily offset, usually
so much that the arrow on the tooltip is disconnected from it.