fix(dropdowns): loosen up auto-placement settings
Allows corner-positioned dropdowns to be auto-positioned in any corner
if the prefered one lacks space for the dropdown to render properly.
Previously, dropdowns could only be auto-positioned on the vertical
axis, meaning that they would remain on the same side of their toggle.
In rare cases, this poses issues for dropdowns that might also lack
horizontal spacing and therefore need to be allowed to flip sides as
well.