feat(GlSingleStat): Format the displayed number based on locale
The numbers displayed by GlSingleStat can visually be a bit
unwieldy when the number gets into the 10's or 100's of thousands.
This MR introduces a new `useDelimiters` boolean prop.
When enabled, the number is formatted based upon the users locale.
The decimal places should be preserved in this process.

Signed-off-by: Robert Hunt <rhunt@gitlab.com>