chore: Component accessibility test issue template
It creates an issue template for documenting the
implementation of a component's accessibility tests.
It includes a section to list the components' states
that should be covered by the tests.
Why an issue template? It saves time by assigning labels,
and epics, and providing a structure to
replicate across issues.