refactor(GlChartTooltip): Tooltip calculates its own position
This refactor modifies the chart tooltip code to make the tooltip
responsible for it's own rendering and position by default.

Currently, many charts have a too wide responsibility as they must
calculate the position on which their tooltips are shown, which causes
code repetition and makes our code prone to errors.

This change only refactors the area chart usage of the tooltip, but a
follow up change can clean up duplicated code from other charts. This
change is backwards compatible, so chart can be updated gradually.