fix(GlAvatarInline): Fix spacing in GlAvatarInlne
Absolute positioning was used previously for this, which necessitated a
lot of manual size management. This resulted in a bug where the parent
component did not have the right width.

This change uses negative right margin on avatars when the avatar-inline
component is collapsed so that the parent element will always have the
same size as the apparent size of the avatars.

This change also corrects the way positive right margins are applied
when the avatars-inline is not collapsed so that the final avatar does
not have unnecessary right margin.