chore(deps): update dependency @gitlab/eslint-plugin to v19.4.0
This ignores most new findings of `unicorn/no-array-callback-reference`,
as they are mostly functions defined in close proximity to their usage,
and are clearly designed as callbacks. The risk of them gaining new
parameters and misbehaving is low.

There are several calls to `Wrapper#find` from `@vue/test-utils` which
cause false positives. Perhaps this rule should be disabled in test
files anyway, but for now, this will do.

This also ignores a reference to `lodash/isString`. This is a function
that is very unlikely to gain a second parameter, by its nature. This is
debatable, but in any case, it only exists in a Vue prop validator,
which only runs in development mode. The risk of breakage is low.

The one call this does change is a reference to a BootstrapVue utility.
It's not a documented export of the package, and so it has no reason to
keep its API as-is.

Finally, this removes some duplication in the deprecated dropdown file
to call an already-defined `filterVisible` utility function.