feat(GlBaseDropdown): Use different positioning strategy to display horizontal content
Most of our dropdowns disclose vertically and may contain dynamically changing content (and height)
so we prefer using a most space strategy, provided by autoPlacement, to avoid flipping
content.

Some of our dropdowns may display content to the sides of the dropdown, and in these cases
a flip strategy is better.