fix(dropdowns): revert clipping changes
Reverts recent changes to dropdowns' clipping behavior. We had attempted
to make dropdowns' positioning smarter, but that ended up caussing
issues with dropdowns that are rendered within a clipping element that
is not `<main>`. Because there isn't an easy solution to address all
issues at once, we've decided to revert those changes and let the
consumers set the proper positioning option when needed.