test: assert that trigger buttons have the focus
Some stories have trigger buttons that are expecteds to have the focus
after they have been actioned with in the `play` functions. This adds
assertions to ensure those buttons do have the focus to stabilize visual
regression tests. The screenshots should only be captured when the
buttons have the focus outline.