refactor(storybook): calculate LinkToSource upfront
We currently use `context.require` to get possible source files for our
stories. The problem: Webpack actually seems to try to parse those
files which errors out when vendoring bootstrap vue.

If we do the mapping logic beforehand and inject it with webpack, we
circumvent the problem because we are not reading the file at all. On
the other hand we are not getting a potential "live reload" if someone
creates a new story, they would need to restart webpack.

But on the plus side, the logic is now more portable if we choose to
swap webpack with vite as the builder.