fix(duochat): Wait for another enter to send on IME
When using IME such as japanese keyboard, users are given the
opportunity to fix their typing before sending the message. That means
formac this would require three enters (pick from list, confirm
selection, and then send) and with Google keyboard IME it requires two
(pick from list, and then send). So we wait for one additional enter to
send the chat prompt

Normally, this is fixed using a form, but since we are adding more
functionality from the slash commands this didn't seem to work on this
context (example of fix using a form:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/147814/diffs)

Changelog: fixed