feat(tailwind): add the custom configuration
This backports the Tailwind configuration we have been iterating on in
the GitLab project. This results in a more or less complete config which
consumer can start leveraging to apply Pajamas-compliant CSS utils.

BREAKING CHANGE: This overrides several Tailwind configuration option.
Consumers that were already using Tailwind with the GitLab UI preset
should keep an eye out for utils that might behave differently.