feat(Charts): use options.toolbox.show to show/hide toolbox
This MR updates the charts component to use the `toolbox.show`
property from options to render the chart toolbox.

Previously the show (boolean) attribute was not being used,
instead a toolbox would be rendered if an object was passed
in specifying the SVG icons to be used for the toolbox.