feat(GlTokenSelector): Add `showAddNewAlways` prop
This MR adds the prop `showAddNewAlways` to
GlTokenSelector which enables the ability for
users to add new tokens regardless if
there are currently results in the list.

This change also updates the dropdown list
to include the user defined token as an actual
navigatable dropdown item in the list.

Additionally there are changes to the logic
to create a SSoT for all components regarding
if a new token should be able to be created or
not.