feat: Add boolean flag for improvedHideHeuristics in Config
The config.js now accepts a new boolean flag. When defined, it sets
a global property using the Vue Global API. The components
GlDisclosureDropdown and GlCollapsibleListbox now use a mixin
which is setting a property, and a computed property, using either
the property, falling back to the global property (secondary).