chore: Fix GlSorting story in production build
Before, the GlSorting story wasn't rendering in production. Bizarrely,
the story literally rendered the HTML `<gl-sorting ... ></gl-sorting>`,
as if Vue is treating it as a custom element. Meanwhile, it rendered
correctly in the development build.

For reasons that aren't clear, changing how the story is defined fixes
this.

Addresses https://gitlab.com/gitlab-org/gitlab-ui/-/issues/2780.