fix: Installation as a dependency
With 8e82e6ec11f987d4097e734f8a76bd5334b67bb1 we broke downstream
projects, since the `postinstall` script tried to run `patch-package` in
that project, which isn't correct.

For instance, see
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/178190.

Now, we only run `patch-package` when we're about to run Storybook,
which is the thing we're patching.