feat(GlTooltip): add custom global container selector
This change allows us to use a custom container
for tooltips instead of <body> by default.
This is important to mitigate layout recalculations
that happen each time we append a new tooltip as
the last element on the <body>.
Using custom container will not cause severe style
recalculations as the <body> element does.