feat(GlDrawer): Remove explicit border
This removes the border for every child element ("unboxing") and instead
leaves it up to each implementation to determine when a border is
needed.

Sections within a drawer that previously had a bottom border may no
longer have one if it was from the component styles. This may be desired
if the content within should align with the change to be "unboxed."
There's no change to padding or margin, so content should still remain
visually separated.

If a border is still needed, utility classes can be used.

The "sidebar" variant of the drawer still has these styles, and so
should not be affected. In any case, it's likely that variant will be
removed in future.